tiangou/requirements.txt
amqp==2.5.1
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
Django==1.11.11
django-cors-headers==3.1.0
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
environment==1.0.0
Flask==1.1.1
Flask-Session==0.3.1
idna==2.8
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
kombu==4.6.4
MarkupSafe==1.1.1
more-itertools==7.2.0
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
pycryptodomex==3.7.2
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.2
qiniu==7.2.6
redis==3.3.8
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
view==0.1
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
